Switch to_bits_spec from int to u32/u64

+++ source/builtin/src/lib.rs
@@ -1340,14 +1340,14 @@ impl_binary_op!(SpecShr, spec_shr, Self, [
 #[cfg(verus_keep_ghost)]
 #[verifier::spec]
 #[rustc_diagnostic_item = "verus::verus_builtin::f32_to_bits"]
-pub fn f32_to_bits(_f: f32) -> int {
+pub fn f32_to_bits(_f: f32) -> u32 {
     unimplemented!()
 }
 
 #[cfg(verus_keep_ghost)]
 #[verifier::spec]
 #[rustc_diagnostic_item = "verus::verus_builtin::f64_to_bits"]
-pub fn f64_to_bits(_f: f64) -> int {
+pub fn f64_to_bits(_f: f64) -> u64 {
     unimplemented!()
 }
